In StanleyDec: Integer --> long

+++ source/libnormaliz/simplex.h
@@ -91,7 +91,7 @@ class SimplexEvaluator {
     Matrix<Integer> RS; // right hand side to hold order vector
     // Matrix<Integer> RSmult; // for multiple right hand sides
     
-    Matrix<Integer>* StanleyMat;
+    Matrix<long>* StanleyMat;
     size_t StanIndex;
     size_t nr_level0_gens; // counts the number of level 0 vectors among the generators
